+++ rt_gene/src/rt_gene/kalman_stabilizer.py
@@ -51,44 +51,44 @@ class Stabilizer(object):
         self.filter = cv2.KalmanFilter(state_num, measure_num, 0)
 
         # Store the state.
-        self.state = np.zeros((state_num, 1), dtype=np.float32)
+        self.state = np.zeros((state_num, 1), dtype=float)
 
         # Store the measurement result.
-        self.measurement = np.array((measure_num, 1), np.float32)
+        self.measurement = np.array((measure_num, 1), float)
 
         # Store the prediction.
-        self.prediction = np.zeros((state_num, 1), np.float32)
+        self.prediction = np.zeros((state_num, 1), float)
 
         # Kalman parameters setup for scalar.
         if self.measure_num == 1:
             self.filter.transitionMatrix = np.array([[1, 1],
-                                                     [0, 1]], np.float32)
+                                                     [0, 1]], float)
 
-            self.filter.measurementMatrix = np.array([[1, 1]], np.float32)
+            self.filter.measurementMatrix = np.array([[1, 1]], float)
 
             self.filter.processNoiseCov = np.array([[1, 0],
-                                                    [0, 1]], np.float32) * cov_process
+                                                    [0, 1]], float) * cov_process
 
             self.filter.measurementNoiseCov = np.array(
-                [[1]], np.float32) * cov_measure
+                [[1]], float) * cov_measure
 
         # Kalman parameters setup for point.
         if self.measure_num == 2:
             self.filter.transitionMatrix = np.array([[1, 0, 1, 0],
                                                      [0, 1, 0, 1],
                                                      [0, 0, 1, 0],
-                                                     [0, 0, 0, 1]], np.float32)
+                                                     [0, 0, 0, 1]], float)
 
             self.filter.measurementMatrix = np.array([[1, 0, 0, 0],
-                                                      [0, 1, 0, 0]], np.float32)
+                                                      [0, 1, 0, 0]], float)
 
             self.filter.processNoiseCov = np.array([[1, 0, 0, 0],
                                                     [0, 1, 0, 0],
                                                     [0, 0, 1, 0],
-                                                    [0, 0, 0, 1]], np.float32) * cov_process
+                                                    [0, 0, 0, 1]], float) * cov_process
 
             self.filter.measurementNoiseCov = np.array([[1, 0],
-                                                        [0, 1]], np.float32) * cov_measure
+                                                        [0, 1]], float) * cov_measure
 
     def update(self, measurement):
         """Update the filter"""
@@ -97,10 +97,10 @@ class Stabilizer(object):
 
         # Get new measurement
         if self.measure_num == 1:
-            self.measurement = np.array([[np.float32(measurement[0])]])
+            self.measurement = np.array([[float(measurement[0])]])
         else:
-            self.measurement = np.array([[np.float32(measurement[0])],
-                                         [np.float32(measurement[1])]])
+            self.measurement = np.array([[float(measurement[0])],
+                                         [float(measurement[1])]])
 
         # Correct according to mesurement
         self.filter.correct(self.measurement)
@@ -112,13 +112,13 @@ class Stabilizer(object):
         """Set new value for processNoiseCov and measurementNoiseCov."""
         if self.measure_num == 1:
             self.filter.processNoiseCov = np.array([[1, 0],
-                                                    [0, 1]], np.float32) * cov_process
+                                                    [0, 1]], float) * cov_process
             self.filter.measurementNoiseCov = np.array(
-                [[1]], np.float32) * cov_measure
+                [[1]], float) * cov_measure
         else:
             self.filter.processNoiseCov = np.array([[1, 0, 0, 0],
                                                     [0, 1, 0, 0],
                                                     [0, 0, 1, 0],
-                                                    [0, 0, 0, 1]], np.float32) * cov_process
+                                                    [0, 0, 0, 1]], float) * cov_process
             self.filter.measurementNoiseCov = np.array([[1, 0],
-                                                        [0, 1]], np.float32) * cov_measure
+                                                        [0, 1]], float) * cov_measure
